+++ docs/changes.rst
@@ -21,6 +21,7 @@ Not yet released.
 * Allow to avoid overwriting approved translations on file upload.
 * Dropped support for some compatibility URL redirects.
 * Added check for ECMAScript template literals.
+* Added option to watch a component.
 
 Weblate 4.1.1
 -------------